+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### dev-master
 * Fix error during WordPress installation
 * Add `WP_STASH_BYPASS` environment variable
+* object-cache.php drop-in now merely requires the actual drop-in from the install folder. This makes it update-safe
 
 #### 3.1.0
 * Implementation of `Debug\ActionLogger`